Updates to 64bit capable. Mostly void * -> long int
Not sure who sent in this patch, but appears to work great


git-svn-id: svn://svn.code.sf.net/p/btonline-btech/code/btmux/trunk@1472 e60b412a-5d77-420b-80c2-6cef64185626

+++ btech/btpr/include/config.h
@@ -22,8 +22,8 @@
 
 #include "debug.h"
 
-typedef int	dbref;
-typedef int	FLAG;
+typedef long    dbref;
+typedef long	FLAG;
 typedef int     POWER;
 
 /* TEST_MALLOC:	Defining this makes a malloc that keeps track of the number
